Graveyard Started & Terrain Added
Created a graveyard fence, as well as a four gravestone prefabs we may replace with better-looking assets later.

In addition, I have organized different gravestone sets that include a mixture of the gravestone prefabs in a row.

Finally, I've sized everything down a little and added a 3D terrain for future convenience.